fixed q-gram types
[git-p4: depot-paths = "//sw/gpgpu/nvbio/main/": change = 18208072]

+++ nvbio/qgram/qgram_inl.h
@@ -53,7 +53,7 @@ void QGramIndexDevice::build(
     qgrams.resize( string_len );
     index.resize( string_len );
 
-    thrust::device_vector<uint32> d_all_qgrams( string_len );
+    thrust::device_vector<uint64> d_all_qgrams( string_len );
 
     // build the list of q-grams
     thrust::transform(
@@ -230,7 +230,7 @@ void QGramSetIndexDevice::build(
     qgrams.resize( n_qgrams );
     index.resize( n_qgrams );
 
-    thrust::device_vector<uint32> d_all_qgrams( n_qgrams );
+    thrust::device_vector<uint64> d_all_qgrams( n_qgrams );
 
     // build the list of q-gram indices
     thrust::transform(
